backend-python/requirements.txt
flask==2.3.3
mysql-connector-python==8.1.0
pandas==2.1.0
requests==2.31.0
